pyproject.toml
[tool.poetry]
name = "streakbot"
version = "0.1.0"
description = ""
authors = ["yangken"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.4"
pyTelegramBotAPI = "^4.5.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
